pyproject.toml
[tool.poetry]
name = "eduplatform"
version = "0.1.0"
description = ""
authors = ["Sherif Ragimov <ragimov700@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "4.2.10"
flake8 = "^7.0.0"
djangorestframework = "^3.14.0"
djoser = "^2.2.2"
django-debug-toolbar = "^4.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
